divide a fat code related canvas delegate
